+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchModel.kt
@@ -54,8 +54,4 @@ data class OverdueSearchModel(
   fun selectedOverdueAppointmentsChanged(selectedOverdueAppointments: Set<UUID>): OverdueSearchModel {
     return copy(selectedOverdueAppointments = selectedOverdueAppointments)
   }
-
-  fun clearSelectedOverdueAppointments(): OverdueSearchModel {
-    return copy(selectedOverdueAppointments = emptySet())
-  }
 }