+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Use overdue appointment selector in overdue search screen
 - Use overdue appointment selector to fetch selected ids in overdue download worker
 - Use overdue appointment selector to fetch selected ids in overdue download format selection dialog
+- Remove unnecessary code in overdue screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-11-8325